show video status correctly
It was always showing the video status on the vide edit form as public even when it was private.  This was due to setting the default value in the form, instead of on the new video model instance used to populate the form.  This change also includes a cucumber test to verify it.